fix: renamed functional module to fn
