+++ src/fn/array/every.ts
