+++ src/fn/array/for-each.ts
