+++ src/fn/array/index.ts
