+++ src/fn/array/map.ts
