+++ src/fn/array/types.ts
