+++ src/fn/base/clamp.spec.ts
