+++ src/fn/base/clamp.ts
