+++ src/fn/base/divide.spec.ts
