+++ src/fn/base/divide.ts
