+++ src/fn/base/exponential.spec.ts
