+++ src/fn/base/exponential.ts
