+++ src/fn/base/linear.spec.ts
