+++ src/fn/base/linear.ts
