+++ src/fn/base/max.ts
