+++ src/fn/base/min.ts
