+++ src/fn/base/multiply.ts
