+++ src/fn/base/nth-root.ts
