+++ src/fn/base/polynomial.ts
