+++ src/fn/base/pow.ts
