+++ src/fn/base/rational.spec.ts
