+++ src/fn/base/rational.ts
