+++ src/fn/base/scale.spec.ts
