+++ src/fn/base/scale.ts
