+++ src/fn/base/scaleBy.spec.ts
