+++ src/fn/base/scaleBy.ts
