+++ src/fn/base/subtract.ts
