+++ src/fn/flow/expand.spec.ts
