+++ src/fn/flow/expand.ts
