+++ src/fn/index.ts
@@ -1,0 +1,6 @@
+/**
+ * This module holds functions and utilities to be used in a
+ * functional programming style.
+ */
+
+export * from './flow';
