+++ src/fn/numbers/quaternions/multiply-with.spec.ts
