+++ src/fn/numbers/quaternions/multiply-with.ts
