+++ src/fn/types/operator-function.ts
