+++ src/numbers/quaternion/operators.ts
@@ -1,5 +1,5 @@
 import { sum } from '../../base';
-import { chain } from '../../functional';
+import { chain } from '../../fn';
 
 import { Quaternion, wxyz } from './quaternion';
